Add "certificate of reg. number" field
Aircraft are registered in Malta and other countries by means of the issuance of a certificate of registration. These certificates are numbered in the case of the Maltese registry and presumably some others, which allows us to identify the registration uniquely as opposed to tail numbers (registration number) which may be reused.